* Fixed more type cast issues.

+++ base/VulkanBuffer.hpp
@@ -99,7 +99,7 @@ namespace vks
 		void copyTo(void* data, VkDeviceSize size)
 		{
 			assert(mapped);
-			memcpy(mapped, data, size);
+			memcpy(mapped, data, (size_t)size);
 		}
 
 		/**
