+++ base/vulkantextoverlay.hpp
@@ -637,7 +637,7 @@ public:
 		renderPassBeginInfo.clearValueCount = 0;
 		renderPassBeginInfo.pClearValues = nullptr;
 
-		for (int32_t i = 0; i < cmdBuffers.size(); ++i)
+		for (int32_t i = 0; i < (int32_t)cmdBuffers.size(); ++i)
 		{
 			renderPassBeginInfo.framebuffer = *frameBuffers[i];
